This is the main rewrite for the iomemry-vsl driver. All files have been rearranged to be closer to the 4.x series. Especially the new 4.x test cases have been implemented and are used to mitigate bugs not yet covered in the 3.x driver series
- bio error handling
- split of too large bios
- hotplug handling
