+++ root/usr/src/iomemory-vsl-3.2.15/kinfo.c
@@ -390,11 +390,11 @@ static int kfio_info_seqf_open(fusion_inode *inode, fusion_file *file)
 static void kfio_info_os_init(void)
 {
     /* Initialize file ops used to handle fixed types. */
-    kfio_set_file_ops_owner(&kfio_info_type_fops, fusion_get_this_module());
+    kfio_set_file_ops_owner(&kfio_info_type_fops, THIS_MODULE);
     kfio_set_file_ops_write_handler(&kfio_info_type_fops, kfio_info_os_type_write);
     kfio_set_file_ops_read_handler (&kfio_info_type_fops, kfio_info_os_type_read);
 
-    kfio_set_file_ops_owner(&kfio_info_seqf_fops, fusion_get_this_module());
+    kfio_set_file_ops_owner(&kfio_info_seqf_fops, THIS_MODULE);
     kfio_set_file_ops_open_handler   (&kfio_info_seqf_fops, kfio_info_seqf_open);
     kfio_set_file_ops_read_handler   (&kfio_info_seqf_fops, kfio_seq_read);
     kfio_set_file_ops_llseek_handler (&kfio_info_seqf_fops, kfio_seq_lseek);
